Incorporated review suggestions
