+++ examples/platform_services/android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.platformservices"
+    package="com.example.flutter"
     android:versionCode="1"
     android:versionName="0.0.1">
 
@@ -10,7 +10,7 @@
 
     <application android:name="io.flutter.app.FlutterApplication" android:label="@string/app_name" >
         <activity
-            android:name="com.example.flutter.ExampleActivity"
+            android:name=".ExampleActivity"
             android:launchMode="singleTop"
             android:theme="@android:style/Theme.Black.NoTitleBar"
             android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
