+++ packages/flutter/lib/services.dart
@@ -13,15 +13,19 @@ library services;
 export 'src/services/asset_bundle.dart';
 export 'src/services/binding.dart';
 export 'src/services/clipboard.dart';
+export 'src/services/codec.dart';
 export 'src/services/haptic_feedback.dart';
 export 'src/services/image_cache.dart';
 export 'src/services/image_decoder.dart';
 export 'src/services/image_provider.dart';
 export 'src/services/image_resolution.dart';
 export 'src/services/image_stream.dart';
+export 'src/services/json_codec.dart';
 export 'src/services/path_provider.dart';
+export 'src/services/platform_channel.dart';
 export 'src/services/platform_messages.dart';
 export 'src/services/raw_keyboard.dart';
+export 'src/services/standard_codec.dart';
 export 'src/services/system_chrome.dart';
 export 'src/services/system_navigator.dart';
 export 'src/services/system_sound.dart';
